lib: modem_key_mgmt: move header to /include/modem/
Move header file to the new /include/modem/ directory.
